refactor: move RefreshRuntimeModule to fix ESLint
