+++ lib/RefreshRuntimeModule.js
@@ -1,8 +1,8 @@
 const RuntimeGlobals = require('webpack/lib/RuntimeGlobals');
 const RuntimeModule = require('webpack/lib/RuntimeModule');
 const Template = require('webpack/lib/Template');
-const { refreshGlobal } = require('../globals');
-const getRefreshGlobal = require('../utils/getRefreshGlobal');
+const { refreshGlobal } = require('./globals');
+const { getRefreshGlobal } = require('./utils');
 
 class ReactRefreshRuntimeModule extends RuntimeModule {
   constructor() {
