+++ lib/index.js
@@ -233,7 +233,7 @@ class ReactRefreshPlugin {
           case 5: {
             const NormalModule = require('webpack/lib/NormalModule');
             const RuntimeGlobals = require('webpack/lib/RuntimeGlobals');
-            const ReactRefreshRuntimeModule = require('./runtime/RefreshRuntimeModule');
+            const ReactRefreshRuntimeModule = require('./RefreshRuntimeModule');
 
             compilation.hooks.additionalTreeRuntimeRequirements.tap(
               this.constructor.name,
